Terminando a api
