+++ src/config/passport.js.js
@@ -1,13 +1,13 @@
 
-let JwtStrategy = require('passport-jwt').Strategy;
-let ExtractJwt = require('passport-jwt').ExtractJwt;
-let User = require('../db/models/users');
-let config = require('../config/config');
+const JwtStrategy = require('passport-jwt').Strategy;
+const ExtractJwt = require('passport-jwt').ExtractJwt;
+const User = require('../db/models/users');
+const config = require('../config/config');
 
 
 // Setup work and export for the JWT passport strategy
 module.exports = function(passport) {
-  let opts = {
+  const opts = {
     jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
     secretOrKey: config.jwt_encryption
   };
@@ -27,47 +27,3 @@ module.exports = function(passport) {
     });
   }));
 };
-
-/*const passport = require('passport');
-const LocalStrategy = require('passport-local').Strategy;
-const passportJWT = require('passport-jwt');
-const JWTStrategy = passportJWT.Strategy;
-const bcrypt = require('bcrypt');
-
-
-
-
-const { secret } = CONFIG.jwt_encryption;
-
-const UserModel = require('../db/models/users');
-
-passport.use(new LocalStrategy({
-  usernameField: email,
-  passwordField: password,  
-}, async (username, password, done) => {
-  try {
-    const userDocument = await UserModel.findOne({username: username}).exec();
-    const passwordsMatch = await bcrypt.compare(password, userDocument.passwordHash);
-
-    if (passwordsMatch) {
-      return done(null, userDocument);
-    } else {
-      return done('Incorrect Username / Password');
-    }
-  } catch (error) {
-    done(error);
-  }
-}));
-
-passport.use(new JWTStrategy({
-    jwtFromRequest: req => req.cookies.jwt,
-    secretOrKey: secret,
-  },
-  (jwtPayload, done) => {
-    if (jwtPayload.expires > Date.now()) {
-      return done('jwt expired');
-    }
-
-    return done(null, jwtPayload);
-  }
-));*/
