+++ src/controller/UserController.js
@@ -1,0 +1,86 @@
+let jwt = require('jsonwebtoken');
+var User = require('../db/models/users');
+let config = require('../config/config');
+
+let UserController = {};
+
+UserController.register = ((req, res) => {
+    if (!req.body.email || !req.body.password) {
+        console.log(req.body.email)
+        res.json({
+            success: false,
+            message: 'Please enter email and password.'
+        });
+    } else {
+        let newUser = new User({
+            email: req.body.email,
+            password: req.body.password,
+            admin: req.body.admin
+        });
+
+        // Attempt to save the user
+        newUser.save(function (err) {
+            if (err) {
+                return res.json({
+                    success: false,
+                    message: 'That email address already exists.'
+                });
+            }
+            res.json({
+                success: true,
+                message: 'Successfully created new user.'
+            });
+        });
+    }
+});
+
+UserController.find = ((req, res) => {
+    console.log("Segundo");
+    User.find().then((User) => {
+        console.log("QUE PORRA")
+		res.send({
+			User
+		});
+	}, (e) => {
+        console.log("DEU ERROS");
+        
+		res.status(400).send(e);
+	})
+})
+
+UserController.auth = ((req, res) => {
+    User.findOne({
+        email: req.body.email
+    }, function (err, user) {
+        if (err) throw err;
+
+        if (!user) {
+            res.send({
+                success: false,
+                message: 'Authentication failed. User not found.'
+            });
+        } else {
+            // Check if password matches
+            user.comparePassword(req.body.password, function (err, isMatch) {
+                if (isMatch && !err) {
+                    // Create token if the password matched and no error was thrown
+                    var token = jwt.sign(user.toJSON(), config.jwt_encryption, {
+                        expiresIn: 10000
+                    });
+                    res.json({
+                        success: true,
+                        message: 'Authentication successfull',
+                        token
+                    });
+                } else {
+                    res.send({
+                        success: false,
+                        message: 'Authentication failed. Passwords did not match.'
+                    });
+                }
+            });
+        }
+    });
+});
+
+module.exports = UserController;
