--- src/index2.js
@@ -1,45 +1,0 @@
-//Velho
-/*
-import http from 'http';
-import express from 'express';
-import cors from 'cors';
-import morgan from 'morgan';
-import bodyParser from 'body-parser';
-import initializeDb from './db';
-import middleware from './middleware';
-import api from './api';
-import config from './config.json';
-
-let app = express();
-//Adicionando porta
-const PORT = 8000;
-app.server = http.createServer(app);
-
-// logger
-app.use(morgan('dev'));
-
-// 3rd party middleware
-app.use(cors({
-	exposedHeaders: config.corsHeaders
-}));
-
-app.use(bodyParser.json({
-	limit : config.bodyLimit
-}));
-
-// connect to db
-initializeDb( db => {
-
-	// internal middleware
-	app.use(middleware({ config, db }));
-
-	// api router
-	app.use('/api', api({ config, db }));
-
-	app.server.listen(process.env.PORT || config.port, () => {
-		console.log(`Started on port ${app.server.address().port}`);
-	});
-});
-
-export default app;
-*/
